requirements.txt
flask
flask-script
flask-login
flask-sqlalchemy
sqlalchemy-migrate
flask-wtf
flask-bcrypt
flask-debugtoolbar
